[Chore] ArticleManager로 파싱 코드 변경
